refactor: refactoring gatsby node utils
