+++ www/gatsby-node-functions/index.js
@@ -1,0 +1,9 @@
+const createPages = require('./createPages');
+const onCreateNode = require('./onCreateNode');
+const onCreateWebpackConfig = require('./onCreateWebpackConfig');
+const createCssUtilityClassNodes = require('./createCssUtilityClassNodes');
+const onCreatePage = require('./onCreatePage');
+
+module.exports = {
+  createPages, onCreateNode, onCreateWebpackConfig, createCssUtilityClassNodes, onCreatePage,
+};
