+++ www/gatsby-node-functions/utils/component-page-tabs.js
