+++ www/gatsby-node-functions/utils/index.js
@@ -1,0 +1,6 @@
+const { retrieveRootFiles } = require('./component-page-tabs');
+const { getThemesSCSSVariables, processComponentSCSSVariables } = require('./theme-variables');
+
+module.exports = {
+  retrieveRootFiles, getThemesSCSSVariables, processComponentSCSSVariables,
+};
